+++ changelog.md
@@ -3,6 +3,7 @@
 ### New / Changed Widgets
 - UZSU widgets support one-time events with "once" property back in time on the current day using "activeToday" property 
 - UZSU widgets dnow disable editing the events until the backend has updated the dict with data from the former edit
+- device uzsuicon supports the "perday" option for interpolation (interpolation with the days events only, not across day borders)
 
 ### Other New Features